requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-mail==1.4.1
firebase-admin==6.5.0
gcloud==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-play-scraper==1.2.6
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.5.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.3
msgpack==1.0.8
oauth2client==4.1.3
orjson==3.9.10
passlib==1.7.4
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.17
psycopg2==2.9.9
py-bcrypt==0.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-jwt==2.0.1
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.32.0.post1
style==1.1.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
